Refactors friendlyDate()
Separates out the step which find the right
translation key for the friendly date from the
stage which translates that key.

In particular, adds a new function,
`friendlyDateTranslationKey()`, which just finds
the right translation key. `friendlyDate()` now
just calls the first function and applies the
translation function from i18next.